Moved Update and delete logic for dialogues into a separate container
