+++ src/RPGTools/DialogueTreeMaker/domain/graphUtil.ts
@@ -18,7 +18,6 @@ export const convertDialoguesToNodes = (
                 label: name,
                 size: 20,
                 color: '#d6a840',
-                type,
             },
         };
     };
